Create Check_BST.py
This is a python code to check whether the given tree is a Binary Search Tree or not. A BST is a binary tree, which means each node has at most two children: a left child and a right child. Nodes are ordered in a way that for each node, all nodes in its left subtree have values less than the node, and all nodes in its right subtree have values greater than the node. This property is known as the "BST property" and stands it out form the general tree data structure.